refactor: Add Clockify namespace
