+++ src/Bot.csproj
@@ -10,7 +10,7 @@
     <ItemGroup>
         <PackageReference Include="AdaptiveCards" Version="2.3.0" />
         <PackageReference Include="AdaptiveCards.Templating" Version="1.1.0" />
-        <PackageReference Include="Clockify.Net" Version="1.2.0" />
+        <PackageReference Include="Clockify.Net" Version="1.11.0" />
         <PackageReference Include="F23.StringSimilarity" Version="3.1.0" />
         <PackageReference Include="FluentDateTime" Version="2.0.0" />
         <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="3.1.1" />
